+++ main/settings.py
@@ -32,7 +32,6 @@ from redbeat import RedBeatScheduler
 
 from main.celery_utils import OffsettingSchedule
 from main.sentry import init_sentry
-from main.utils import get_float
 from openapi.settings_spectacular import open_spectacular_settings
 
 VERSION = "0.118.1"
@@ -41,6 +40,36 @@ log = logging.getLogger()
 
 # set log level on cssutils - should be fairly high or it will log messages for Outlook-specific styling
 cssutils.log.setLevel(logging.CRITICAL)
+
+
+class EnvironmentVariableParseException(ImproperlyConfigured):
+    """Environment variable was not parsed correctly"""
+
+
+def get_float(name, default):
+    """
+    Get an environment variable as an int.
+
+    Args:
+        name (str): An environment variable name
+        default (float): The default value to use if the environment variable doesn't exist.
+
+    Returns:
+        float:
+            The environment variable value parsed as an float
+    """  # noqa: E501
+    value = os.environ.get(name)
+    if value is None:
+        return default
+
+    try:
+        parsed_value = float(value)
+    except ValueError as ex:
+        msg = f"Expected value in {name}={value} to be a float"
+        raise EnvironmentVariableParseException(msg) from ex
+
+    return parsed_value
+
 
 ENVIRONMENT = get_string(
     name="MITX_ONLINE_ENVIRONMENT",
